// types.ts
export interface Account {
  id: string;
  email: string;
  nickname: string;
  color: string;
  tistoryId?: string;
  tistoryPassword?: string;
  tistoryBlogUrl?: string;
  wordpressUsername?: string;
  wordpressPassword?: string;
  wordpressUrl?: string;
  blogspotUsername?: string;
  blogspotPassword?: string;
  blogspotUrl?: string;
  naverId?: string;
  naverPassword?: string;
  naverBlogUrl?: string;
  chromeProfile?: string;
}